+++ src/Larasub.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace Err0r\Larasub;
-
-class Larasub {}